Added a Temp Hotspot
I really wanted the Set button to trigger a temporary hotspot.  I am 99% certain the code in this will work.  I tested it without the timeout and it was fine.  It does need the timeout fixed...

But there's a huge problem... the buttons appear to have no debounce so... One press results in a bunch of results...  I've left my attempts at debounce in the code but... I'll come back to it later.  For now, the buttons are still unfunctional.