mDNS works, +/- Brightness, Readme update
I pushed a bunch of little updates... Well, here's a big one!  I had to hack around the Brightness area a lot to get that working (it's independent of the config and temporary for now).  I probably removed some stuff too because with all the hacking I did, the code itself may not know what triggered the brightness change.  Lol.
 The mDNS also works fully.  The device should restart after changing the name due to the mDNS as well.  And the Readme has a bunch of re-formatting, some big blocks split into sub-sections... Whew.  I think I'm done for now.  If this were for sale, I might buy it!